generator/requirements.txt
pydantic>=2.0
pyyaml>=6.0
rich>=13.0
